+++ tests/maskers/test_image.py
